@@ -9,7 +9,7 @@ using System.Collections.Generic;
 
 namespace SpiritKing.Components.Nodes;
 
-public class Scene : Interfaces.IDrawable, Interfaces.IUpdateable
+public abstract class Scene : Interfaces.IDrawable, Interfaces.IUpdateable
 {
     public List<Interfaces.IDrawable> DrawableNodes = new();
     public List<Interfaces.IUpdateable> UpdateableNodes = new();
@@ -20,7 +20,7 @@ public class Scene : Interfaces.IDrawable, Interfaces.IUpdateable
 
     public OrthographicCamera Camera { get; set; }
 
-    public virtual event Action<Scene> SceneSwitched;
+    public event Action<Scene>? SceneSwitched;
 
     public int DrawOrder { get; } = 0;
 
@@ -28,14 +28,15 @@ public class Scene : Interfaces.IDrawable, Interfaces.IUpdateable
 
     public Color BackgroundColor { get; set; }
 
-    public bool Enabled => true;
+    public bool Enabled { get; protected set; } = true;
 
     public int UpdateOrder => 1;
 
-    public bool Visible => true;
+    public bool Visible { get; protected set; } = true;
 
     public Texture2D _logo;
 
+    protected SaveDataController SaveDataController { get; private set; } = new SaveDataController();
     public Scene(Game game)
     {
         Game = game;
@@ -49,7 +50,10 @@ public class Scene : Interfaces.IDrawable, Interfaces.IUpdateable
     {
         foreach (var node in UpdateableNodes)
         {
-            node.Update(gameTime);
+            if (node.Enabled)
+            {
+                node.Update(gameTime);
+            }
         }
     }
 
@@ -58,10 +62,13 @@ public class Scene : Interfaces.IDrawable, Interfaces.IUpdateable
         Game.GraphicsDevice.Clear(BackgroundColor);
         var transformMatrix = Camera.GetViewMatrix();
         spriteBatch.Begin(transformMatrix: transformMatrix);
-
+        
         foreach (var node in DrawableNodes)
         {
-            node.Draw(gameTime, spriteBatch);
+            if (node.Visible)
+            {
+                node.Draw(gameTime, spriteBatch);
+            }
         }
         spriteBatch.End();
     }
@@ -91,4 +98,7 @@ public class Scene : Interfaces.IDrawable, Interfaces.IUpdateable
         }
         MusicController.Unload();
     }
+
+    //TODO: this is stupid, at least change name.
+    protected virtual void OnSceneSwitched(Scene scene) => SceneSwitched?.Invoke(scene);
 }
