@@ -6,37 +6,83 @@ using MonoGame.Extended.Particles.Modifiers.Interpolators;
 using MonoGame.Extended.Particles.Profiles;
 using SpiritKing.Components.Nodes;
 using SpiritKing.Controllers;
+using SpiritKing.Data;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
+using static SpiritKing.Data.SettingsData;
 
 namespace SpiritKing.Scenes;
 
 public class TitleScreenScene : Scene
 {
     private readonly Label TitleText;
-    private readonly MenuButton StartBtn;
-    private readonly MenuButton SettingsBtn;
-    private readonly MenuButton ExitBtn;
-
+    private List<MenuButton> SaveSlotBtns = [];
+    private List<MenuButton> MainMenuBtns = [];
     private readonly MenuController _menuController;
-
     readonly ParticleController _bottomOfScreenParticles;
-    public override event Action<Scene> SceneSwitched;
-
+    private SettingsData _settingsData;
     public TitleScreenScene(Game game) : base(game)
     {
-        TitleText = new Label(game, "Posessed Will", Vector2.Zero, Color.DarkRed);
-        StartBtn = new MenuButton(game, new Point(0, 0), new Point(350, 0), "Start", 0);
-        SettingsBtn = new MenuButton(game, new Point(0, 0), new Point(350, 0), "Settings", 0);
-        ExitBtn = new MenuButton(game, new Point(0, 0), new Point(350, 0), "Exit", 0);
-        _menuController = new MenuController(new MenuButton[] { StartBtn, SettingsBtn, ExitBtn });
-
-        StartBtn.Action = StartClicked;
-        SettingsBtn.Action = SettingsClicked;
-        ExitBtn.Action = ExitClicked;
-        StartBtn.Highlighted = true;
-        BackgroundColor = Color.Black;
+        _settingsData = SaveDataController.LoadGameSettingsAsync().Result;
         var screen = game.GraphicsDevice.Viewport;
+
+        if (_settingsData == null)
+        {
+            SaveDataController.SaveGameSettingsAsync(new()).Wait();
+            _settingsData = new();
+        }
+        TitleText = new Label(game, "Posessed Will", Vector2.Zero, Color.DarkRed);
+        TitleText.Position = new Vector2((screen.Width - TitleText.STRING_SIZE.X) / 2, (screen.Height - TitleText.STRING_SIZE.Y) / 12);
+        var bottomOfTitleText = (int)(TitleText.Position.Y + TitleText.STRING_SIZE.Y);
+
+        foreach (var save in _settingsData.SaveSlots)
+        {
+            SaveSlotBtns.Add(new SaveSlotMenuButton(Game, Point.Zero, new(400, 0), save)
+            {
+                Action = ExistingSaveClicked
+            });
+
+        }
+        SaveSlotBtns.Add(new(Game, Point.Zero, new(400, 0), "Back")
+        {
+            Action = DisplayMainMenu
+        });
+
+        if (_settingsData.CurrentSave == null)
+        {
+            MainMenuBtns.Add(new MenuButton(game, Point.Zero, new Point(400, 0), "New Game")
+            {
+                Action = NewGameClicked,
+                Highlighted = true
+            });
+        }
+        else
+        {
+            MainMenuBtns.Add(new MenuButton(game, Point.Zero, new Point(400, 0), "Continue")
+            {
+                Action = ContinueBtnClicked,
+                Highlighted = true
+            });
+
+            MainMenuBtns.Add(new MenuButton(game, Point.Zero, new Point(400, 0), "Load Game")
+            {
+                Action = LoadGameClicked,
+                ChildButtons = SaveSlotBtns
+            });
+        }
+
+        MainMenuBtns.Add(new MenuButton(game, Point.Zero, new Point(400, 0), "Settings")
+        {
+            Action = SettingsClicked
+        });
+
+        MainMenuBtns.Add(new MenuButton(game, Point.Zero, new Point(400, 0), "Exit")
+        {
+            Action = ExitClicked
+        });
+
+        BackgroundColor = Color.Black;
         _bottomOfScreenParticles = new ParticleController(game.GraphicsDevice, new Vector2(screen.Width / 2, screen.Height));
 
         _bottomOfScreenParticles.AddEmitter(
@@ -74,15 +120,30 @@ public class TitleScreenScene : Scene
                     }
             });
 
+        _menuController = new MenuController(new Point(0, bottomOfTitleText), new Vector2(screen.Width, screen.Height - bottomOfTitleText));
+
         AddNode(_bottomOfScreenParticles);
         AddNode(_menuController);
         AddNode(TitleText);
-        MusicController.PlaySong(MusicController.OrganTheme, true);
 
-        TitleText.Position = new Vector2((screen.Width - TitleText.STRING_SIZE.X) / 2, (screen.Height - TitleText.STRING_SIZE.Y) / 8);
-        StartBtn.Position = new Point((screen.Width - StartBtn.Size.X) / 2, 2 * (screen.Height - StartBtn.Size.Y) / 5);
-        SettingsBtn.Position = new Point((screen.Width - SettingsBtn.Size.X) / 2, 3 * (screen.Height - SettingsBtn.Size.Y) / 5);
-        ExitBtn.Position = new Point((screen.Width - ExitBtn.Size.X) / 2, 4 * (screen.Height - ExitBtn.Size.Y) / 5);
+        _menuController.SetMenuButtons(MainMenuBtns);
+
+        _menuController.UpdateMenuButtonsAlignement();
+    }
+
+    private void ExistingSaveClicked(MenuButton button)
+    {
+        if (button is SaveSlotMenuButton saveSlot)
+        {
+            var gameData = SaveDataController.LoadGameSaveAsync(saveSlot.FileName).Result;
+            base.OnSceneSwitched(new GameScene(Game, gameData));
+        }
+    }
+
+    private void DisplayMainMenu(MenuButton button)
+    {
+        _menuController.SetMenuButtons(MainMenuBtns);
+        _menuController.UpdateMenuButtonsAlignement();
     }
 
     public override void Update(GameTime gameTime)
@@ -98,17 +159,40 @@ public class TitleScreenScene : Scene
         TitleText.Dispose();
     }
 
-    private void StartClicked()
+    private void NewGameClicked(MenuButton caller)
     {
-        SceneSwitched?.Invoke(new GameScene(Game));
+        //TODO
     }
 
-    private void SettingsClicked()
+    private void ContinueBtnClicked(MenuButton caller)
+    {
+        var gameData = SaveDataController.LoadGameSaveAsync(_settingsData.CurrentSave.FileName).Result;
+        base.OnSceneSwitched(new GameScene(Game, gameData));
+    }
+
+    private void LoadGameClicked(MenuButton caller)
+    {
+        _menuController.SetMenuButtons(caller.ChildButtons);
+        _menuController.UpdateMenuButtonsAlignement();
+    }
+
+    private void SettingsClicked(MenuButton caller)
     {
         Debug.WriteLine("Settings clicked.");
+        var saves = SaveDataController.GetSaveFileNames();
+
+        if (saves.Count <= 0)
+        {
+            if (_settingsData.SaveSlots.Count == saves.Count)
+            {
+                _settingsData.SaveSlots.Add(new SaveSlot(true, "Occupied Save", 0, "Save1.dat"));
+                _settingsData.CurrentSave = _settingsData.SaveSlots[0];
+            }
+        }
+        SaveDataController.SaveGameSettingsAsync(_settingsData).Wait();
     }
 
-    private void ExitClicked()
+    private void ExitClicked(MenuButton caller)
     {
         Dispose();
         Game.Exit();
